requirements.txt
beautifulsoup4
lxml
matplotlib
numpy
psycopg2-binary
requests
tqdm